# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a API-Jira server.",
        "version": "0.0.1",
        "title": "Jira-API",
        "termsOfService": ""
    },
    "host": "",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/get/jira/object/includeAttributObject": {
            "get": {
                "tags": [
                    "Get Jira Objects"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "objectSchemaName",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "IBVSD"
                    },
                    {
                        "name": "objectTypeName",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "Customer"
                    },
                    {
                        "name": "attribute",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "Sender"
                    },
                    {
                        "name": "value",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "herbert.tung"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/set/jira/issue/createIssue": {
            "post": {
                "tags": [
                    "Create Jira Issue"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "createIssue",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/issueAttribute"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/set/jira/issue/updateIssue": {
            "post": {
                "tags": [
                    "Create Jira Issue"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "updateIssue",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/issueAttributeUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
...
        "/get/jira/issue/issue": {
            "get": {
                "tags": [
                    "Get Jira Issue"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "issueId",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "TOCP-85"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/IBVSD/gotEmail": {
            "post": {
                "tags": [
                    "IBVSD"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/IBVSDGotEmailDef"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
        "/IBVSD/updateEmailToIssue": {
            "post": {
                "tags": [
                    "IBVSD"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/IBVSDUpdateIssueDef"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "issueAttribute": {
            "type": "object",
            "properties": {
                "createIssue": {
                    "type": "object",
                    "properties": {
                        "project": {
                            "type": "object",
                            "properties": {
                                "key": {
                                    "type": "string",
                                    "enum": "AM"
                                }
                            }
                        },
                        "issuetype": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "enum": "Computer Asset"
                                }
                            }
                        },
                        "summary": {
                            "type": "string",
                            "enum": "Created by API"
                        },
                        "customfield_12317": {
                            "type": "string",
                            "enum": "77790211"
                        },
                        "customfield_12318": {
                            "type": "string",
                            "enum": "910040611"
                        },
                        "customfield_12320": {
                            "type": "string",
                            "enum": "D12-345678"
                        },
                        "customfield_12321": {
                            "type": "string",
                            "enum": "1809-008841"
                        },
                        "customfield_12322": {
                            "type": "string",
                            "enum": "FGL1839263S"
                        },
                        "customfield_12327": {
                            "type": "string",
                            "enum": "MACROVIEW,RENT"
                        },
                        "customfield_12328": {
                            "type": "string",
                            "enum": "SI quotation no.: QN/2018/07245/TH"
                        },
                        "customfield_12329": {
                            "type": "string",
                            "enum": "Macroview"
                        },
                        "customfield_12330": {
                            "type": "string",
                            "enum": "2011-10-03"
                        },
                        "customfield_12331": {
                            "type": "string",
                            "enum": "2011-10-03"
                        },
                        "customfield_12332": {
                            "type": "object",
                            "properties": {
                                "value": {
                                    "type": "string",
                                    "enum": "ROUTER"
                                }
                            }
                        },
                        "customfield_12333": {
                            "type": "object",
                            "properties": {
                                "value": {
                                    "type": "string",
                                    "enum": "CISCO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "issueAttributeUpdate": {
            "type": "object",
            "properties": {
                "updateIssue": {
                    "example": {
                        "issueId": "IBVSD-9",
                        "fields": {
                            "customfield_12020": "TEST"
                        }
                    }
                }
            }
        },
...
        "IBVSDGotEmailDef": {
            "type": "object",
            "properties": {
                "info": {
                    "example": {
                        "issueKey": "IBVSD-9",
                        "to": "herbert.tung@hgc.com.hk",
                        "cc": "testing@hgc.com.hk",
                        "subject": "testing open issue",
                        "body": "Email body"
                    }
                }
            }
        },
        "IBVSDUpdateIssueDef": {
            "type": "object",
            "properties": {
                "issueKey": {
                    "example": "IBVSD-9"
                },
                "email": {
                    "example": {
                        "CustomerTo": "herbert.tung@hgc.com.hk",
                        "CustomerCc": "herbert.tung@hgc.com.hk",
                        "CustomerSubject": "[IBVSD-1] Testing",
                        "CustomerBody": "Email body"
                    }
                }
            }
        }
    }
}